backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
openai
python-jose
passlib
python-dotenv
sqlalchemy
alembic
psycopg2-binary
numpy
pandas
matplotlib
seaborn
asyncpg
'pydantic[email]'
bcrypt